Fix unintentional disabling of console.debug
Previously, console.debug was overridden to only print to the console if debugLogs
in config.js was set to true. This causes unintended side effects for any code that
uses this module. To fix, console.debug is no longer overridden and the debug logs
are turned on or off by a condition check for each case.